Stripped out the OpenMP SIMD macros as they're not worth the effort.
In dense scenarios, the compiler already autovectorizes properly (based on some
godbolt testing), so the SIMD pragmas have no effect.

Admittedly, they are useful for forcing vectorization in sparse cases, as the
compiler can't prove that the indices don't alias. But in those cases, the
execution time is likely dominated by memory lookups when the indices don't
refer to the same cache line... which they won't, in typical sparse data.

In light of this, we remove these macros to reduce the risk of incorrectness
and the amount of testing we have to do. We don't even OpenMP tests anyway.